Editor: Unified App Config/Play/Publish into single panel in SideBar.
