+++ editor/js/libs/app/index.html
@@ -22,10 +22,8 @@
 
 			import * as THREE from './js/three.module.js';
 			import { APP } from './js/app.js';
-			import { VRButton } from './js/VRButton.js';
 
 			window.THREE = THREE; // Used by APP Scripts.
-			window.VRButton = VRButton; // Used by APP Scripts.
 
 			var loader = new THREE.FileLoader();
 			loader.load( 'app.json', function ( text ) {
